+++ constants/frameworks.go
@@ -1,0 +1,10 @@
+package constants
+
+// Frameworks defines the supported frameworks within the application
+// Supported frameworks:
+// Generic
+// Node.js
+var Frameworks = map[string]string{
+	"VERSION":      "generic",
+	"package.json": "node",
+}