+++ main.go
@@ -2,9 +2,14 @@ package main
 
 import (
 	"flag"
+
+	"github.com/fatih/color"
+	"github.com/oshalygin/go-tag/constants"
+	"github.com/oshalygin/go-tag/services"
 )
 
 var versionFlagPassed bool
+var directory = "."
 
 func init() {
 	flag.BoolVar(&versionFlagPassed, "version", false, "Print the version of the application")
@@ -16,5 +21,13 @@ func main() {
 		printVersion()
 	}
 
-	// services.Tag("0.0.3")
+	framework, err := services.DetermineFrameworks(directory, constants.Frameworks)
+
+	if err != nil {
+		color.Red("No framework files matched")
+		return
+	}
+
+	color.Green(framework)
+
 }